some more work on fluids

+++ HumanSpaceProgram/Assets/HSP.ResourceFlow/FlowEdge.cs
@@ -10,9 +10,9 @@ namespace HSP.ResourceFlow
         public readonly FlowNode end1;
         public readonly FlowNode end2;
 
-        public float Volume { get; }
+        public double Volume { get; }
 
-        public FlowEdge( FlowNode end1, FlowNode end2, float volume )
+        public FlowEdge( FlowNode end1, FlowNode end2, double volume )
         {
             this.end1 = end1;
             this.end2 = end2;
